update, select, delete with seeder

+++ database/migrations/2021_08_07_191052_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration
             $table->string('password');
             $table->rememberToken();
             $table->timestamps();
-            $table->integer('profession_id'); 
+            $table->unsignedBigInteger('profession_id'); 
             $table->foreign('profession_id')->references('id')->on('professions');
         });
     }
